[refactor] 회원가입 시 이메일 전송 링크를 통한 인증 Application 클래스 -> UserService 클래스로 통합 및 삭제
