+++ src/main/java/org/board/exercise_board/user/domain/repository/UserRepository.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
   Optional<User> findByLoginId(String loginId);
-  Boolean existsByEmail(String email);
+  boolean existsByEmail(String email);
+  boolean existsByLoginId(String loginId);
 }
